+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 * Replaced `Pronto.gem_names` with `Pronto::GemNames.new.to_a`.
 * [#116](https://github.com/mmozuras/pronto/pull/116): improved GitHub formatter error output.
 * [#123](https://github.com/mmozuras/pronto/pull/126): added runner attribute to message initialization.
+* Runner expects to receive patches/commit via `initialize(patches, commit)`, instead of `run(patches, commit)`.
 
 ### Bugs fixed